Drop namespace
